gs: Restructure LNS messages package
