+++ pkg/gatewayserver/io/basicstationlns/messages/messages.go
@@ -18,19 +18,15 @@ import (
 	"encoding/json"
 	"fmt"
 	"strings"
-	"time"
 
 	"go.thethings.network/lorawan-stack/pkg/band"
 	"go.thethings.network/lorawan-stack/pkg/basicstation"
-	"go.thethings.network/lorawan-stack/pkg/encoding/lorawan"
 	"go.thethings.network/lorawan-stack/pkg/errors"
 	"go.thethings.network/lorawan-stack/pkg/frequencyplans"
-	"go.thethings.network/lorawan-stack/pkg/ttnpb"
 )
 
 var (
-	errFrequencyPlan      = errors.Define("frequency_plan", "invalid frequency plan")
-	errJoinRequestMessage = errors.Define("join_request_message", "invalid join-request message received")
+	errFrequencyPlan = errors.Define("frequency_plan", "invalid frequency plan")
 )
 
 // Definition of message types
@@ -55,12 +51,6 @@ const (
 	configHardwareSpecPrefix            = "sx1301"
 	configHardwareSpecNoOfConcentrators = "1"
 )
-
-// DataRates encodes the available datarates of the channel plan for the Station in the format below
-// [0] -> SF (Spreading Factor; Range: 7...12 for LoRa, 0 for FSK)
-// [1] -> BW (Bandwidth; 125/250/500 for LoRa, ignored for FSK)
-// [2] -> DNONLY (Downlink Only; 1 = true, 0 = false)
-type DataRates [16][3]int
 
 // DiscoverQuery contains the unique identifier of the gateway.
 // This message is sent by the gateway.
@@ -196,97 +186,4 @@ func GetRouterConfig(fp frequencyplans.FrequencyPlan, isProd bool) (RouterConfig
 	// TODO: Harmonize this with sx1301_conf from https://github.com/TheThingsNetwork/lorawan-stack/pull/276
 	cfg.SX1301Config = SX1301Config{}
 	return cfg, nil
-}
-
-// UpInfo provides additional metadata on each upstream message.
-type UpInfo struct {
-	RxTime  int64   `json:"rxtime"`
-	RCtx    int64   `json:"rtcx"`
-	XTime   int64   `json:"xtime"`
-	GPSTime int64   `json:"gpstime"`
-	RSSI    float32 `json:"rssi"`
-	SNR     float32 `json:"snr"`
-}
-
-// RadioMetaData is a the metadata that is received as part of all upstream messages (except Tx Confirmation).
-type RadioMetaData struct {
-	DataRate  int    `json:"DR"`
-	Frequency uint64 `json:"Freq"`
-	UpInfo    UpInfo `json:"upinfo"`
-}
-
-// JoinRequest is the LoRaWAN Join Request message
-type JoinRequest struct {
-	MHdr          uint             `json:"MHdr"`
-	JoinEUI       basicstation.EUI `json:"JoinEui"`
-	DevEUI        basicstation.EUI `json:"DevEui"`
-	DevNonce      uint             `json:"DevNonce"`
-	MIC           int32            `json:"MIC"`
-	RefTime       float64          `json:"RefTime"`
-	RadioMetaData RadioMetaData
-}
-
-// MarshalJSON implements json.Marshaler.
-// TODO: Make MarshalJSON() messages generic.
-func (req JoinRequest) MarshalJSON() ([]byte, error) {
-	type Alias JoinRequest
-	return json.Marshal(struct {
-		Type string `json:"msgtype"`
-		Alias
-	}{
-		Type:  TypeUpstreamJoinRequest,
-		Alias: Alias(req),
-	})
-}
-
-// ToUplinkMessage extracts fields from the basic station Join Request "jreq" message and converts them into an UplinkMessage for the network server.
-func (req *JoinRequest) ToUplinkMessage(ids ttnpb.GatewayIdentifiers, bandID string) (ttnpb.UplinkMessage, error) {
-	up := ttnpb.UplinkMessage{}
-	up.ReceivedAt = time.Now()
-
-	parsedMHDR := ttnpb.MHDR{}
-	err := lorawan.UnmarshalMHDR([]byte{byte(req.MHdr)}, &parsedMHDR)
-	if err != nil {
-		return ttnpb.UplinkMessage{}, errJoinRequestMessage.WithCause(err)
-	}
-
-	micBytes, err := getInt32AsByteSlice(req.MIC)
-	if err != nil {
-		return ttnpb.UplinkMessage{}, errJoinRequestMessage.WithCause(err)
-	}
-	up.Payload = &ttnpb.Message{
-		MHDR: parsedMHDR,
-		MIC:  micBytes,
-		Payload: &ttnpb.Message_JoinRequestPayload{JoinRequestPayload: &ttnpb.JoinRequestPayload{
-			JoinEUI:  req.JoinEUI.EUI64,
-			DevEUI:   req.DevEUI.EUI64,
-			DevNonce: [2]byte{byte(req.DevNonce), byte(req.DevNonce >> 8)},
-		}},
-	}
-
-	up.RawPayload, err = lorawan.MarshalMessage(*up.Payload)
-	if err != nil {
-		return ttnpb.UplinkMessage{}, errJoinRequestMessage.WithCause(err)
-	}
-
-	rxTime := time.Unix(req.RadioMetaData.UpInfo.RxTime, 0)
-	rxMetadata := &ttnpb.RxMetadata{
-		GatewayIdentifiers: ids,
-		Time:               &rxTime,
-		Timestamp:          uint32(req.RadioMetaData.UpInfo.XTime & 0xFFFFFFFF),
-		RSSI:               req.RadioMetaData.UpInfo.RSSI,
-		SNR:                req.RadioMetaData.UpInfo.SNR,
-	}
-	up.RxMetadata = append(up.RxMetadata, rxMetadata)
-
-	loraDR, err := getDataRateFromIndex(bandID, req.RadioMetaData.DataRate)
-	if err != nil {
-		return ttnpb.UplinkMessage{}, errJoinRequestMessage.WithCause(err)
-	}
-	up.Settings = ttnpb.TxSettings{
-		Frequency:     req.RadioMetaData.Frequency,
-		DataRateIndex: (ttnpb.DataRateIndex)(req.RadioMetaData.DataRate),
-		DataRate:      loraDR,
-	}
-	return up, nil
 }
